[YS-531] refactor: 모바일 내 정보 수정 페이지 Participant 디렉토리 분리
