+++ src/app/user/profile/mobile/edit/[infoType]/components/Participant/index.ts
@@ -1,0 +1,11 @@
+import AddressEditForm from './AddressEditForm/AddressEditForm';
+import ContactEmailEditForm from './ContactEmailEditForm/ContactEmailEditForm';
+import MatchTypeEditForm from './MatchTypeEditForm/MatchTypeEditForm';
+import NameEditForm from './NameEditForm/NameEditForm';
+
+export const Participant = {
+  ContactEmail: ContactEmailEditForm,
+  Name: NameEditForm,
+  Address: AddressEditForm,
+  MatchType: MatchTypeEditForm,
+};
